Added confirm dialog during logout

+++ lib/components/app_drawer.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:glancefrontend/screens/home_screen.dart';
+import 'package:glancefrontend/screens/jobcards/jobcards_main_screen.dart';
 import 'package:glancefrontend/screens/login_screen.dart';
 import 'package:glancefrontend/screens/settings_screen.dart';
 import 'package:glancefrontend/services/claim_data_service.dart';
@@ -42,17 +43,21 @@ class AppDrawer extends StatelessWidget {
             leading: const Icon(Icons.home),
             title: const Text('Home'),
             onTap: () {
-              Navigator.pushAndRemoveUntil(context, MaterialPageRoute(builder: (ctx) => const HomeScreen()), (route) => false);
+              Navigator.pushAndRemoveUntil(context, MaterialPageRoute(builder: (_) => const HomeScreen()), (route) => false);
+            },
+          ),
+          ListTile(
+            leading: const Icon(Icons.card_membership),
+            title: const Text('Job Cards'),
+            onTap: () {
+              Navigator.push(context, MaterialPageRoute(builder: (_) => const JobCardsScreen()));
             },
           ),
           ListTile(
             leading: const Icon(Icons.settings),
             title: const Text('Settings '),
             onTap: () {
-              Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                      builder: (context) => const SettingsScreen()));
+              Navigator.push(context, MaterialPageRoute(builder: (_) => const SettingsScreen()));
             },
           ),
           const Divider(color: Colors.black),
@@ -60,9 +65,31 @@ class AppDrawer extends StatelessWidget {
             leading: const Icon(Icons.logout),
             title: const Text('Logout '),
             onTap: () {
-              LocalStorage.deleteAll();
-              Navigator.push(context,
-                  MaterialPageRoute(builder: (context) => const LoginScreen()));
+              showDialog(
+                context: context,
+                builder: (context) {
+                  return AlertDialog(
+                    title: const Text('Logout'),
+                    content: const Text('Are you sure you want to logout?'),
+                    actions: [
+                      TextButton(
+                        child: const Text('No'),
+                        onPressed: () {
+                          Navigator.of(context).pop(false);
+                        },
+                      ),
+                      TextButton(
+                        child: const Text('Yes'),
+                        onPressed: () {
+                          LocalStorage.deleteAll();
+                          Navigator.of(context).pop(true);
+                          Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (_) => const LoginScreen()));
+                        },
+                      ),
+                    ],
+                  );
+                }
+              );
             },
           ),
         ],
